Login Script to force MBAM encryption at login
This script will check if MBAM client is installed and if the  OS drive is encrypted. if not the script ask to the reboot if they do not. the script give them 8 hrs and then force reboot. but the script will warn the user at 4 hrs ,30 minutes, 15 minutes and 5 minutes before the restart